Add zhpeq_reserve_next() to support iov changes
The iov handling implemented in the zhpe provider was
arcane and expensive. This operator is one of the changes that
allowed it to be simplified by providing a mechanism to try
to reserve another queue slot after an existing one. The iov
code tries to gang multiple operations into a single doorbell
using this.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>