Fix EBUSY in mmaptest
Request ZMMU for mmap was not being marked as individual and
went down the wrong path when freed. Also changed the
zhpe_mmap_unmap() API not to accept addr and length since those
are problematic. Race condtions exist in the test, but I hope
those can be addressed when the commit operation is fixed.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>